+++ game/game_libs/game_shared/weapons/generichitscanweapon.cpp
@@ -99,8 +99,9 @@ bool CGenericHitscanWeapon::InvokeWithAttackMode(
 		using namespace EventConstructor;
 		CEventConstructor event;
 
-		event << Flags(DefaultEventFlags()) << Invoker(m_pPlayer->edict()) << EventIndex(eventID)
-			  << IntParam1(m_pPlayer->random_seed) << BoolParam1(m_iClip == 0) << FloatParam1(GetInaccuracy());
+		event << Flags(DefaultEventFlags()) << Invoker(m_pPlayer->edict())
+			  << EventIndex(static_cast<unsigned short>(eventID)) << IntParam1(m_pPlayer->random_seed)
+			  << BoolParam1(m_iClip == 0) << FloatParam1(GetInaccuracy());
 
 		event.Send();
 	}
