+++ game/game_libs/pm_shared/pm_debug.c
@@ -73,9 +73,9 @@ void PM_ShowClipBox(void)
 		pmove->player_maxs[pmove->usehull],
 		org,
 		pmove->server ? 132 : 0,
-		0.1);
+		0.1f);
 
-	PM_ParticleLine(org, org, pmove->server ? 132 : 0, 0.1, 5.0);
+	PM_ParticleLine(org, org, pmove->server ? 132 : 0, 0.1f, 5.0f);
 /*
 	{
 		int i;
@@ -118,7 +118,7 @@ void PM_ParticleLine(vec3_t start, vec3_t end, int pcolor, float life, float ver
 		for ( i = 0; i < 3; i++ )
 			curpos[i] = start[i] + curdist * diff[i];
 
-		pmove->PM_Particle(curpos, pcolor, life, 0, vert);
+		pmove->PM_Particle(curpos, pcolor, life, 0, (int)vert);
 		curdist += linestep;
 	}
 }
@@ -289,7 +289,7 @@ void PM_ViewEntity(void)
 
 	VectorCopy(pmove->origin, origin);
 
-	fup = 0.5 * (pmove->player_mins[pmove->usehull][2] + pmove->player_maxs[pmove->usehull][2]);
+	fup = 0.5f * (pmove->player_mins[pmove->usehull][2] + pmove->player_maxs[pmove->usehull][2]);
 	fup += pmove->view_ofs[2];
 	fup -= 4;
